feature: Chess layer and printing the board
