--- src/boardgame/Board.java
@@ -1,31 +1,0 @@
-package boardgame;
-
-public class Board {
-
-    private int rows;
-    private int columns;
-    private Piece[][] pieces;
-
-    public int getRows() {
-        return rows;
-    }
-
-    public void setRows(int rows) {
-        this.rows = rows;
-    }
-
-    public int getColumns() {
-        return columns;
-    }
-
-    public void setColumns(int columns) {
-        this.columns = columns;
-    }
-
-    public Board(int rows, int columns) {
-        this.rows = rows;
-        this.columns = columns;
-        this.pieces = new Piece[rows][columns];
-    }
-
-}
