+++ src/br/com/henriqueaguiiar/boardgame/Position.java
@@ -1,4 +1,4 @@
-package boardgame;
+package br.com.henriqueaguiiar.boardgame;
 
 public class Position {
     private int row;
